Added the project files and schematics for the board
Note: This addition and update attempts to mostly preserve the original design when the project was handed in for grading in December 2022, but also makes a few necessary changes.

Summary:
This revision, Revision B, corrects 2 bugs, makes some minor improvements, and considerably improves the readability and tidiness of the schematics so that they can be uploaded. Additionally, some traces were rerouted and cleaned up for neatness. For context, the picture of the working prototype in the README is the original board, marked “Rev A.”

Bugs Fixed:
The first bug fixed was the pin choices for two of the ESP32 pins that control RANGE_SW_1 and RANGE_SW_2. Originally, pins 34 and 35 were routed to these switches, but they had to be changed to pins 27 and 14, respectively.

The second bug fixed was that the calibration switch was accidently wired so that it would never register a button press. This was fixed by connecting the CALIBRATE line on the schematic with pin 2 on the switch, instead of pin 3.

Additions:
The addition includes adding a fast-blow fuse on the battery for safety.

Miscellaneous Changes:
The miscellaneous changes include considerably improving the readability and neatness of the schematics; removing unnecessary component pads (which were placed on the power lines in case I needed more filtering than what I did end up needing); and neatening up several of the traces on the PCB.